refactor(routes): move resolve_flag HTTP + MCP endpoints into routes/resolve_flag
Splits the DELETE /routines/{id}/flags/{filename} handler and the MCP
resolve_flag tool into their own src/routes/resolve_flag/ folder, following
the template in src/routes/CONTRIBUTING.md and the create_flag/list_flags
precedent.
