+++ src/routes/mcp_lock_tests.rs
@@ -4,7 +4,6 @@
 )]
 
 use super::*;
-use crate::routes::mcp::mcp_types::CreateFlagInput;
 
 fn make_handler() -> MoadimMcp {
     MoadimMcp::new(
@@ -46,24 +45,6 @@ impl Drop for TempHome {
         unsafe {
             std::env::remove_var("MOADIM_HOME_OVERRIDE");
         }
-    }
-}
-
-fn make_create_routine_req() -> crate::routines::CreateRoutineRequest {
-    crate::routines::CreateRoutineRequest {
-        model: None,
-        schedule: "@daily".into(),
-        title: "Mcp Routine".into(),
-        agent: "claude".into(),
-        prompt: "p".into(),
-        goal: None,
-        repositories: vec![],
-        machines: vec![crate::machine::current_machine()],
-        enabled: true,
-        ttl_secs: None,
-        max_runtime_secs: None,
-        tags: vec![],
-        env: std::collections::HashMap::new(),
     }
 }
 
@@ -332,75 +313,4 @@ fn unlock_routines_returns_error_when_set_lock_fails() {
         std::env::remove_var("MOADIM_HOME_OVERRIDE");
     }
     let _ = std::fs::remove_dir_all(&dir);
-}
-
-// ── flag tools ─────────────────────────────────────────────────────────────────
-
-/// Extract and parse the JSON text out of a tool result's first content item.
-fn result_json(result: &CallToolResult) -> serde_json::Value {
-    let text = match &result.content[0] {
-        rmcp::model::ContentBlock::Text(block) => block.text.clone(),
-        _ => panic!("expected text content"),
-    };
-    serde_json::from_str(&text).unwrap()
-}
-
-#[test]
-fn resolve_flag_not_found_is_error() {
-    use rmcp::handler::server::wrapper::Parameters;
-    let handler = make_handler();
-    let result = handler
-        .resolve_flag(Parameters(ResolveFlagInput {
-            id: "no-such".into(),
-            filename: "bug-1.md".into(),
-        }))
-        .unwrap();
-    assert!(result.is_error.unwrap_or(false));
-}
-
-#[test]
-fn create_list_resolve_flag_lifecycle() {
-    use rmcp::handler::server::wrapper::Parameters;
-    let _home = TempHome::set();
-    let routines = crate::routines::new_store();
-    let handler = MoadimMcp::new(routines, crate::paths::routines_dir(), 0, test_shutdown());
-
-    let created = handler
-        .create_routine(Parameters(make_create_routine_req()))
-        .unwrap();
-    let id = result_json(&created)["id"].as_str().unwrap().to_string();
-
-    // create_flag
-    let result = handler
-        .create_flag(Parameters(CreateFlagInput {
-            id: id.clone(),
-            r#type: "bug".into(),
-            description: "broken thing".into(),
-            scope: "general".into(),
-        }))
-        .unwrap();
-    assert!(!result.is_error.unwrap_or(false));
-    let flag = result_json(&result);
-    assert_eq!(flag["type"], "bug");
-    let filename = flag["filename"].as_str().unwrap().to_string();
-
-    // list_flags
-    let result = handler
-        .list_flags(Parameters(IdInput { id: id.clone() }))
-        .unwrap();
-    assert!(!result.is_error.unwrap_or(false));
-    let flags = result_json(&result);
-    assert_eq!(flags.as_array().unwrap().len(), 1);
-
-    // resolve_flag
-    let result = handler
-        .resolve_flag(Parameters(ResolveFlagInput {
-            id: id.clone(),
-            filename,
-        }))
-        .unwrap();
-    assert!(!result.is_error.unwrap_or(false));
-
-    let result = handler.list_flags(Parameters(IdInput { id })).unwrap();
-    assert_eq!(result_json(&result).as_array().unwrap().len(), 0);
 }
