+++ src/routes/resolve_flag/mod.rs
@@ -1,0 +1,12 @@
+//! Resolve flag: shared logic (`logic.rs`), HTTP handler (`http.rs`), and MCP tool (`mcp.rs`,
+//! declared from `routes::mcp` so it can reach `MoadimMcp`'s private state).
+
+pub(crate) mod logic;
+
+#[path = "http.rs"]
+mod http;
+#[allow(
+    unused_imports,
+    reason = "utoipa's OpenApi derive resolves this hidden __path_resolve_flag type via crate::routes::resolve_flag::__path_resolve_flag, generated by #[utoipa::path] on the re-exported handler below"
+)]
+pub use http::{__path_resolve_flag, resolve_flag};
